requirements.txt
openai==1.30.1
requests==2.31.0
PyGithub==2.3.0
